# api-docs/swagger.yaml
swagger: "2.0"
info:
  description: "Dissertation project - MSc Computer Science."
  version: "1.0.0"
  title: "Diabetes Backend API"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "ec20107@qmul.ac.uk"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:4000"
basePath: "/api/v1"
...
schemes:
  - "https"
  - "http"
paths:
  /medications:
    get:
      tags:
        - "medication"
      summary: "Get the list of medications"
      description: "Get the list of medications registered in the system. If the name query parameter is supplied, it will perform a lookup."
      produces:
        - "application/json"
      parameters:
        - name: "name"
          in: "query"
          description: "The medication name"
          required: false
          type: "string"
      responses:
        "200":
          description: "List of medications"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Medication"
  /medications/{medicationId}:
    get:
      tags:
        - "medication"
      summary: "Get a single medication result based on an ID"
      produces:
        - "application/json"
      parameters:
        - name: "medicationId"
          in: "path"
          description: "ID of medication to return"
          required: true
          type: "string"
      responses: 
        "200":
          description: "A single medication"
          schema: 
            $ref: "#/definitions/Medication"
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  Medication:
    type: "object"
    properties:
      id:
        type: "string"
        example: "92be6243-8e85-46b1-aa96-c9948ebbed99"
      name:
        type: "string"
        example: "Exenatide"
      brandNames:
        type: "array"
        items:
          type: "string"
          example:
            - "Byetta"
            - "Bydureon"
      description:
        type: "string"
        example: "Binds to, and activates, the GLP-1 receptor to increase insulin secretion, suppresses glucagon secretion, and slows gastric emptying."
      drugAdministrationRoute:
        type: "string"
        example: "Injection"
    required:
      - "id"
      - "name"
      - "brandNames"
      - "description"
      - "drugAdministrationRoute"
...